% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{artificialHydrograph}
\alias{artificialHydrograph}
\title{Artificial Hydrograph}
\usage{
artificialHydrograph(step.s = 3600, from = "2015-01-01",
  to = "2015-01-10")
}
\arguments{
\item{step.s}{time step in seconds}

\item{from}{first day as character string in format "yyyy-mm-dd"}

\item{to}{last day as character string in format "yyyy-mm-dd"}
}
\value{
data frame with columns \emph{DateTime} and \emph{values}
}
\description{
Generates an artificial hydrograph
}
