% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseValidation.R
\name{checkForOverlappingTimeSequences}
\alias{checkForOverlappingTimeSequences}
\title{Check for Overlapping Time Sequences}
\usage{
checkForOverlappingTimeSequences(
  dataFrame,
  main = "Overlapping time sequences in hydraulic data"
)
}
\arguments{
\item{dataFrame}{frame with timestamps in first column}

\item{main}{plot title, passed to \code{\link{plotSequenceEvents}}}
}
\value{
\code{TRUE} if there are overlapping sequences, else \code{FALSE}
}
\description{
Check for Overlapping Time Sequences
}
