% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{dataFrameToXts}
\alias{dataFrameToXts}
\title{Convert Time Series Data Frame to xts Object}
\usage{
dataFrameToXts(
  dataFrame,
  timeColumn = names(dataFrame)[posixColumnAtPosition(dataFrame)[1]]
)
}
\arguments{
\item{dataFrame}{data frame containing a timestamp column}

\item{timeColumn}{name or number of time column. Default: first POSIXt-column
in \emph{dataFrame}}
}
\value{
xts object with timestamps taken from timestamp column. Non-numeric
  columns are removed.
}
\description{
Convert Time Series Data Frame to xts Object
}
