% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{demoGroupByInterval}
\alias{demoGroupByInterval}
\title{Demonstrate hsGroupByInterval}
\usage{
demoGroupByInterval(df = hsExampleTSeries(60),
  step = kwb.datetime::minTimeStep(df[, 1]), to_pdf = TRUE)
}
\arguments{
\item{df}{data frame containing data to be used for the demonstration}

\item{step}{time step in seconds}

\item{to_pdf}{if \code{TRUE} the output goes into a PDF file}
}
\description{
Creates a plot demonstrating the effect of offset1 and offset2 in function
hsGroupByInterval
}
