% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{firstTimestamp}
\alias{firstTimestamp}
\title{First Timestamp in Data Frame}
\usage{
firstTimestamp(x)
}
\arguments{
\item{x}{data frame containing a date/time column}
}
\description{
First Timestamp in Data Frame
}
