% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{hsDaylightSaving}
\alias{hsDaylightSaving}
\title{Find Days of Daylight Saving}
\usage{
hsDaylightSaving(year.first = NULL, year.last = NULL)
}
\arguments{
\item{year.first}{first year of which daylight saving dates are to be
calculated}

\item{year.last}{last year of which daylight saving dates are to be
calculated}
}
\value{
data frame with columns \code{begST} (begin of summer time) and
  \code{endST} (end of summer time)
}
\description{
Find Days of Daylight Saving
}
