% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsDbTablePlotXY}
\alias{hsDbTablePlotXY}
\title{X-Y Plot of Two Database Table Fields}
\usage{
hsDbTablePlotXY(strDb, strTable, strX, strY, strPdfFile = NULL)
}
\arguments{
\item{strDb}{full path to MS Access database file (*.mdb)}

\item{strTable}{name of table in database}

\item{strX}{name of table field containing the "x" values}

\item{strY}{name of table field containing the "y" values}

\item{strPdfFile}{full path to PDF file to be generated. If omitted, the plot
is shown on the screen.}
}
\value{
If the output device is a pdf file the result of the dev.off()
  command is returned.
}
\description{
Plots the values of two database fields (= columns) against each other and
prints the plot into a PDF file.
}
\examples{
\dontrun{
## Plot CSB vs. timestamp values from table "tbl_STA_CAL" in the
## example database into a window on the screen
hsDbTablePlotXY(kwb.db::xmdb(), "tbl_Qua", "myDateTime", "CSB")

## Set paths to a test PDF file
pdf_file <- file.path(tempdir(), "ex_hsDbTablePlotXY.pdf")

## Plot CSBf vs. timestamp values from the same table into a PDF file.
hsDbTablePlotXY(kwb.db::xmdb(), "tbl_Qua", "myDateTime", "CSBf", pdf_file)

## Open PDF file in PDF viewer
kwb.utils::hsShowPdf(pdf_file)
}

}
