% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsDelNaRowsOrCols}
\alias{hsDelNaRowsOrCols}
\title{Delete Rows or Columns Containing only NA}
\usage{
hsDelNaRowsOrCols(df, rows = TRUE, drop = FALSE)
}
\arguments{
\item{df}{data frame}

\item{rows}{if TRUE, rows that only contain NAs are deleted, else columns.}

\item{drop}{if TRUE and only one row/column remains this row/column is returned
in forms of a vector instead as a data frame.}
}
\description{
Delete Rows or Columns Containing only NA
}
