% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsExampleTSeries}
\alias{hsExampleTSeries}
\title{Data Frame with Example Time Series}
\usage{
hsExampleTSeries(step)
}
\arguments{
\item{step}{time step in seconds}
}
\value{
data frame with columns \emph{t} (timestamp) and \emph{y} (sinus values)
}
\description{
Example time series
}
