% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFillUp.R
\name{hsFillUp}
\alias{hsFillUp}
\title{Fill-up with Missing Timestamps}
\usage{
hsFillUp(
  tseries,
  tsField = names(tseries)[posixColumnAtPosition(tseries)[1]],
  step_s = 60,
  forceStep = TRUE,
  limits = NULL,
  interpol = TRUE,
  includeOrig = TRUE,
  default = NA,
  dbg = FALSE
)
}
\arguments{
\item{tseries}{data frame representing a time-series of values}

\item{tsField}{Name of timestamp column in tseries. Default: name of first
available POSIXt-column}

\item{step_s}{Time step in seconds that shall lie between consecutive
timestamps. Non-existing timestamps are generated.}

\item{forceStep}{If TRUE, only timestamps that are multiples of the time step
given in step_s are put into the result data frame}

\item{limits}{data frame or matrix with exactly two columns containing first
and last timestamp of intervals for which timestamps are to be generated.}

\item{interpol}{if TRUE, the returned data frame will for each value column
contain a corresponding column containing interpolated values}

\item{includeOrig}{if TRUE and interpol = TRUE the original columns will
remain in the output in addition to the columns containing all (original
plus interpolated) values.}

\item{default}{value to be used if there are not enough (at least to) non-NA
values to be used for interpolation. Default: NA}

\item{dbg}{If TRUE, debug messages are shown}
}
\value{
Returns a data.frame
}
\description{
Based on the time series given in "tseries" a new time series is generated in
which timestamps that are lacking in tseries are included. Optionally missing
values are generated by interpolation between existing values.
}
