% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsFilterPeriod}
\alias{hsFilterPeriod}
\title{Filter Rows Within Time Period}
\usage{
hsFilterPeriod(tSeries, minDate, maxDate, tsField, maxIncluded = FALSE,
  dbg = FALSE)
}
\arguments{
\item{tSeries}{data.frame containing time-series data.}

\item{minDate}{lower boundary of time interval to be selected, as either of
POSIXt-object, Date object or string in ISO format: yyyy-mm-dd}

\item{maxDate}{upper boundary of time interval to be selected, as either of
POSIXt-object, Date object or string in ISO format: yyyy-mm-dd}

\item{tsField}{name of time stamp column in \emph{tseries}.}

\item{maxIncluded}{if TRUE, \emph{maxDate} is included in the time interval
to be filtered for, otherwise it is not included (see Details).}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\value{
Returns \emph{tSeries}, reduced to rows representing a time within
  the selected time interval between \emph{minDate} and \emph{maxDate}
}
\description{
Filters a data.frame containing time series data for the time interval
between \emph{minDate} and \emph{maxDate}.
}
\details{
\emph{minDate} and \emph{maxDate} must be given as character string in ISO
format: \dQuote{yyyy-mm-dd [HH:MM:SS]}, where the brackets indicate that time
information is optional. The lower boundary is always included in the time
interval to be selected, whereas the upper boundary is only included if
\emph{maxIncluded} is TRUE and otherwise excluded. With \emph{maxIncluded} =
FALSE it is easy to select whole months or whole years by setting
\emph{maxDate} to the first day of the next month/year.
}
