% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsFilterRowsWithValuesInColumns}
\alias{hsFilterRowsWithValuesInColumns}
\title{Filter for Rows with Given Values in Given Columns}
\usage{
hsFilterRowsWithValuesInColumns(dframe, columnValuePairs)
}
\arguments{
\item{dframe}{data frame}

\item{columnValuePairs}{list of elements each of which defines a filter
criterion in the form <column-name> = <value>}
}
\value{
data frame containing those rows of \emph{dframe} that comply with
  all of the filter criteria defined in \emph{columnValuePairs}
}
\description{
Filter for rows that have specified values in specified columns
}
