% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsGroupBy2Fields}
\alias{hsGroupBy2Fields}
\title{Two-field Grouping of a Data Frame}
\usage{
hsGroupBy2Fields(frmData, strValField, strField1, strField2,
  boolDesc1 = FALSE, boolDesc2 = FALSE)
}
\arguments{
\item{frmData}{data.frame containing data to be grouped.}

\item{strValField}{name of value field (= column in frmData)}

\item{strField1}{name of first field to be grouped by}

\item{strField2}{name of second field to be grouped by}

\item{boolDesc1}{if TRUE, columns in result matrix will be ordered according
to decreasing values of field1}

\item{boolDesc2}{if TRUE, rows in result matrix will be ordered according to
decreasing values of field2}
}
\value{
matrix with as many rows as there are distinct values in field1 and
  as many columns as there are distinct values in field2 of the input
  data.frame. The matrix contains the sum of values in the specified value
  field of the input data.frame that belong to the corresponding combination
  of values in field1/field2
}
\description{
Groups data by values in two columns creating a matrix with as many rows as
there are distinct values in field1 and as many columns as there are distinct
values in field2. The matrix contains the sum of values in the specified
value field that belong to the corresponding value combination of
field1/field2
}
