% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsMfRows}
\alias{hsMfRows}
\title{Needed Rows for mfrow}
\usage{
hsMfRows(nPlots, nPlotsPerRow)
}
\arguments{
\item{nPlots}{number of total plots}

\item{nPlotsPerRow}{number of plots per row}
}
\value{
Number of rows needed to place all the plots.
}
\description{
Number of rows needed to plot <nPlots> in a grid with <plotCols> plots per
row.
}
