% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{hsMkTimestamps}
\alias{hsMkTimestamps}
\title{Sequence of Timestamps}
\usage{
hsMkTimestamps(from, to, step.s = 60, mdb, tbl, dbg = FALSE)
}
\arguments{
\item{from}{first timestamp in ISO-Syntax: yyyy-mm-dd [HH:MM:SS] where the
part in brackets is optional.}

\item{to}{last timestamp in ISO-Syntax: yyyy-mm-dd [HH:MM:SS] where the part
in brackets is optional.}

\item{step.s}{time step between the timestamps in seconds.}

\item{mdb}{Optional. Full path to MS Access database file (*.mdb).}

\item{tbl}{Optional. Name of table to be created in \emph{mdb}. If no name is
given a name of the type tbl\emph{from}_to_\emph{to}_\emph{step.s}_s is
created. If a table of given name exists, a non-existing name is generated
first, so existing tables will not be overwritten.}

\item{dbg}{if TRUE, debug messages are shown.}
}
\value{
Returns vecotor of timestamps if \emph{mdb} is missing or nothing if
  timestamp table has been generated in database given in \emph{mdb}.
}
\description{
Creates timestamps between first timestamp \emph{from} and \emph{to} with a
distance of \emph{step.s} seconds between the timestamps. If \emph{mdb} is
given, the timestamps are written to a database table in which the timestamp
field is the primary key.
}
\examples{
\dontrun{
## Write timestamps of January 2011 with five minutes step into example db.
hsMkTimestamps("2011-01-01", "2011-02-01", 300, kwb.db::xmdb())
}

## Output:
# Timestamps have been written to table
#   'tblTimestamps_2011_01_01_to_2011_02_01_300s' in
# 'C:/Users/hsonne/Documents/R/win-library/2.14/
#   kwb.base/extdata/RExKwbBase.mdb'.
# Timestamp field has been set as primary key.

}
