% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsNiceLabels}
\alias{hsNiceLabels}
\title{Nice Labels}
\usage{
hsNiceLabels(label, labelstep = NULL, labelpos = NULL, mindist = 1, offset = 0)
}
\arguments{
\item{label}{see \code{\link[kwb.plot]{niceLabels}}}

\item{labelstep}{see \code{\link[kwb.plot]{niceLabels}}}

\item{labelpos}{see \code{\link[kwb.plot]{niceLabels}}}

\item{mindist}{see \code{\link[kwb.plot]{niceLabels}}}

\item{offset}{see \code{\link[kwb.plot]{niceLabels}}}
}
\description{
Generates a nice vector of labels by suppressing labels at certain positions.
Please use kwb.plot::niceLabels instead
}
