% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsPdfDev}
\alias{hsPdfDev}
\title{ID(s) of PDF Device(s)}
\usage{
hsPdfDev(all = FALSE)
}
\arguments{
\item{all}{if TRUE, the ids of all pdf devices are returned, if FALSE, only
the id of the first pdf device}
}
\value{
ID of first (\emph{all} == FALSE) or IDs of all opened pdf devices,
  as e.g. returned by \code{\link{dev.list}}
}
\description{
Returns the IDs of the first or all opened pdf device(s).
}
