% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsPlot}
\alias{hsPlot}
\title{Plot to Specific Device}
\usage{
hsPlot(dev, plotFun = graphics::plot, args)
}
\arguments{
\item{dev}{device id}

\item{plotFun}{plot-function to be used, e.g. \code{plot}, \code{barplot},
\code{points}, \code{lines}, ...}

\item{args}{list of arguments to be passed to the function given in
\emph{plotFun}}
}
\value{
This function returns whtat the plot function given in \code{plotFun}
  returns
}
\description{
Generic function for redirecting a plot-command to a specific device.
}
