% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{hsST2WT}
\alias{hsST2WT}
\title{Convert Summer Time to Winter Time}
\usage{
hsST2WT(tstamps, dbg = FALSE)
}
\arguments{
\item{tstamps}{timestamps in summer time}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
time series, shifted to winter time (timezone is set to "UTC")
}
\description{
Conversion of time series in summer time to time series in winter time
}
