% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBase.R
\name{hsWait}
\alias{hsWait}
\title{Wait for Specified Time}
\usage{
hsWait(secs = 1)
}
\arguments{
\item{secs}{number of seconds to wait}
}
\description{
Waits for the specified number of seconds.
}
