% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{intervalWidthToSeconds}
\alias{intervalWidthToSeconds}
\title{Interval Width to Seconds}
\usage{
intervalWidthToSeconds(intervalWidth)
}
\arguments{
\item{intervalWidth}{character string starting with numeric characters 0-9
and ending with either of "d" (days), "h" (hours) or "m" (minutes)}
}
\description{
Interval Width to Seconds
}
