% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{lastTimestamp}
\alias{lastTimestamp}
\title{Last Timestamp in Data Frame}
\usage{
lastTimestamp(x)
}
\arguments{
\item{x}{data frame containing a date/time column}
}
\description{
Last Timestamp in Data Frame
}
