% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseValidation.R
\name{plotSequenceEvents}
\alias{plotSequenceEvents}
\title{Plot Sequence Events}
\usage{
plotSequenceEvents(timestamps, sequences,
  main = "Overlapping time sequences in hydraulic data",
  language = "de")
}
\arguments{
\item{timestamps}{vector of timestamps}

\item{sequences}{data frame as returned by
\code{\link{getOverlappingTimeSequences}} with attribute "sequenceNumber"}

\item{main}{plot title}

\item{language}{"de" (German) or something else (English)}
}
\description{
Plot Sequence Events
}
