% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDateTime.R
\name{selectTimeIntervalDays}
\alias{selectTimeIntervalDays}
\title{Select Time Interval Days}
\usage{
selectTimeIntervalDays(dat, days = 7,
  firstDay = as.character(as.Date(lastDay) - days),
  lastDay = substr(utils::tail(dat$DateTime, 1), 1, 10), dbg = TRUE)
}
\arguments{
\item{dat}{data frame with column "DateTime"}

\item{days}{number of "last" days to select}

\item{firstDay}{first day as text, in "yyyy-mm-dd" format}

\item{lastDay}{last day as text, in "yyyy-mm-dd" format}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Select Time Interval Days
}
